AK: Granularize AK/Math.h
No behavior change.
