+++ AK/Math/Cbrt.h
@@ -1,0 +1,57 @@
+/*
+ * Copyright (c) 2021-2026, Leon Albrecht <leon2002.la@gmail.com>.
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#pragma once
+
+#include <AK/Concepts.h>
+
+#include <AK/Math/Macros.h>
+
+namespace AK {
+
+template<FloatingPoint T>
+constexpr T cbrt(T x)
+{
+    CONSTEXPR_STATE(cbrt, x);
+    if (__builtin_isinf(x) || x == 0)
+        return x;
+    if (x < 0)
+        return -cbrt(-x);
+
+    T r = x;
+    T ex = 0;
+
+    while (r < 0.125l) {
+        r *= 8;
+        ex--;
+    }
+    while (r > 1.0l) {
+        r *= 0.125l;
+        ex++;
+    }
+
+    r = (-0.46946116l * r + 1.072302l) * r + 0.3812513l;
+
+    while (ex < 0) {
+        r *= 0.5l;
+        ex++;
+    }
+    while (ex > 0) {
+        r *= 2.0l;
+        ex--;
+    }
+
+    r = (2.0l / 3.0l) * r + (1.0l / 3.0l) * x / (r * r);
+    r = (2.0l / 3.0l) * r + (1.0l / 3.0l) * x / (r * r);
+    r = (2.0l / 3.0l) * r + (1.0l / 3.0l) * x / (r * r);
+    r = (2.0l / 3.0l) * r + (1.0l / 3.0l) * x / (r * r);
+
+    return r;
+}
+
+#include <AK/Math/UndefMacros.h>
+
+}
