+++ AK/Math/Constants.h
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2021-2026, Leon Albrecht <leon2002.la@gmail.com>.
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#pragma once
+
+#include <AK/Concepts.h>
+
+namespace AK {
+
+template<FloatingPoint T>
+constexpr T NaN = __builtin_nan("");
+template<FloatingPoint T>
+constexpr T Infinity = __builtin_huge_vall();
+template<FloatingPoint T>
+constexpr T Pi = 3.141592653589793238462643383279502884L;
+template<FloatingPoint T>
+constexpr T E = 2.718281828459045235360287471352662498L;
+template<FloatingPoint T>
+constexpr T Sqrt2 = 1.414213562373095048801688724209698079L;
+template<FloatingPoint T>
+constexpr T Sqrt1_2 = 0.707106781186547524400844362104849039L;
+
+template<FloatingPoint T>
+constexpr T L2_10 = 3.321928094887362347870319429489390175864L;
+template<FloatingPoint T>
+constexpr T L2_E = 1.442695040888963407359924681001892137L;
+
+}
