+++ AK/Math/UndefMacros.h
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2021-2026, Leon Albrecht <leon2002.la@gmail.com>.
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#undef CALL_BUILTIN
+#undef CONSTEXPR_STATE
+#undef AARCH64_INSTRUCTION
