+++ Meta/check-style.py
@@ -33,6 +33,8 @@ LICENSE_HEADER_CHECK_EXCLUDES = {
 # We check that "#pragma once" is present
 PRAGMA_ONCE_STRING = '#pragma once'
 PRAGMA_ONCE_CHECK_EXCLUDES = {
+    'AK/Math/Macros.h',
+    'AK/Math/UndefMacros.h',
     'Userland/Libraries/LibC/assert.h',
     'Ladybird/AppKit/System/Detail/Header.h',
     'Ladybird/AppKit/System/Detail/Footer.h',
